requirements.txt
datasets==2.16.1
huggingface-hub==0.25.1
json5==0.9.14
numpy==1.26.4
pandas==2.2.3
Pillow==10.0.0
python-dateutil==2.8.2
pytorch-lightning==0.9.0
pytorch-nlp==0.5.0
scikit-learn==1.3.0
scipy==1.14.1
tokenizers==0.19.1
tqdm==4.66.1
transformers==4.43.3